pyproject.toml
[project]
name = "ai-sales"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular[sidecar]>=0.28.0",
    "gunicorn>=23.0.0",
    "imap-tools>=1.11.0",
    "ipykernel>=6.29.5",
    "langchain[openai]>=0.3.26",
    "langgraph>=0.5.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "twilio>=9.6.1",
]
